+++ pkg/domain/collector.go
@@ -5,10 +5,11 @@ import libapi "github.com/cloudogu/k8s-support-archive-lib/api/v1"
 type CollectorType string
 
 const (
-	CollectorTypeLog        CollectorType = "Logs"
-	CollectorTypeVolumeInfo CollectorType = "VolumeInfo"
-	CollectorTypeNodeInfo   CollectorType = "NodeInfo"
-	CollectorTypSecret      CollectorType = "Resources/Secrets"
+	CollectorTypeLog         CollectorType = "Logs"
+	CollectorTypeVolumeInfo  CollectorType = "VolumeInfo"
+	CollectorTypeNodeInfo    CollectorType = "NodeInfo"
+	CollectorTypSecret       CollectorType = "Resources/Secrets"
+	CollectorTypeSystemState CollectorType = "Resources/SystemState"
 )
 
 func (c CollectorType) GetConditionType() string {
@@ -21,11 +22,13 @@ func (c CollectorType) GetConditionType() string {
 		return libapi.ConditionNodeInfoFetched
 	case CollectorTypSecret:
 		return libapi.ConditionSecretsFetched
+	case CollectorTypeSystemState:
+		return libapi.ConditionSystemStateFetched
 	default:
 		return ""
 	}
 }
 
 type CollectorUnionDataType interface {
-	PodLog | VolumeInfo | LabeledSample | SecretYaml
+	PodLog | VolumeInfo | LabeledSample | SecretYaml | UnstructuredResource
 }